requirements.txt
channels==4.1.0
Django==5.0.6
nfcpy==1.0.4
pyscard==2.0.10
